refactor: split Actions
