+++ .dumi/theme/builtins/Previewer/CodePreviewer.tsx
@@ -1,40 +1,21 @@
 /* eslint-disable react-hooks-extra/no-direct-set-state-in-use-effect */
 import React, { useEffect, useRef, useState } from 'react';
-import { LinkOutlined, ThunderboltOutlined, UpOutlined } from '@ant-design/icons';
+import { UpOutlined } from '@ant-design/icons';
 import type { Project } from '@stackblitz/sdk';
-import stackblitzSdk from '@stackblitz/sdk';
-import { Badge, Flex, Tooltip } from 'antd';
+import { Badge, Tooltip } from 'antd';
 import { createStyles, css } from 'antd-style';
 import classNames from 'classnames';
 import { FormattedMessage, useLiveDemo, useSiteData } from 'dumi';
-import LZString from 'lz-string';
 
 import useLocation from '../../../hooks/useLocation';
 import BrowserFrame from '../../common/BrowserFrame';
 import ClientOnly from '../../common/ClientOnly';
 import CodePreview from '../../common/CodePreview';
 import EditButton from '../../common/EditButton';
-import CodePenIcon from '../../icons/CodePenIcon';
-import CodeSandboxIcon from '../../icons/CodeSandboxIcon';
-import ExternalLinkIcon from '../../icons/ExternalLinkIcon';
 import DemoContext from '../../slots/DemoContext';
 import SiteContext from '../../slots/SiteContext';
-import CodeBlockButton from './CodeBlockButton';
 import type { AntdPreviewerProps } from '.';
-
-function compress(string: string): string {
-  return LZString.compressToBase64(string)
-    .replace(/\+/g, '-') // Convert '+' to '-'
-    .replace(/\//g, '_') // Convert '/' to '_'
-    .replace(/=+$/, ''); // Remove ending '='
-}
-
-const track = ({ type, demo }: { type: string; demo: string }) => {
-  if (!window.gtag) {
-    return;
-  }
-  window.gtag('event', 'demo', { event_category: type, event_label: demo });
-};
+import Actions from './Actions';
 
 const useStyle = createStyles(({ token }) => {
   const { borderRadius } = token;
@@ -105,13 +86,11 @@ const CodePreviewer: React.FC<AntdPreviewerProps> = (props) => {
     containerRef: demoContainer as React.RefObject<HTMLElement>,
   });
   const anchorRef = useRef<HTMLAnchorElement>(null);
-  const codeSandboxIconRef = useRef<HTMLFormElement>(null);
-  const codepenIconRef = useRef<HTMLFormElement>(null);
   const [codeExpand, setCodeExpand] = useState<boolean>(false);
   const { theme } = React.use(SiteContext);
 
   const { hash, pathname, search } = location;
-  const docsOnlineUrl = `https://ant.design${pathname}${search}#${asset.id}`;
+  const docsOnlineUrl = `https://ant.design${pathname ?? ''}${search ?? ''}#${asset.id}`;
 
   const [showOnlineUrl, setShowOnlineUrl] = useState<boolean>(false);
 
@@ -121,11 +100,6 @@ const CodePreviewer: React.FC<AntdPreviewerProps> = (props) => {
       process.env.NODE_ENV === 'development' || regexp.test(window.location.hostname),
     );
   }, []);
-
-  const handleCodeExpand = (demo: string) => {
-    setCodeExpand((prev) => !prev);
-    track({ type: 'expand', demo });
-  };
 
   useEffect(() => {
     if (asset.id === hash.slice(1)) {
@@ -373,112 +347,24 @@ createRoot(document.getElementById('container')).render(<Demo />);
               dangerouslySetInnerHTML={{ __html: description }}
             />
           )}
-          <Flex wrap gap="middle" className="code-box-actions">
-            {showOnlineUrl && (
-              <Tooltip title={<FormattedMessage id="app.demo.online" />}>
-                <a
-                  className="code-box-code-action"
-                  target="_blank"
-                  rel="noreferrer"
-                  href={docsOnlineUrl}
-                >
-                  <LinkOutlined aria-label="open in new tab" className="code-box-online" />
-                </a>
-              </Tooltip>
-            )}
-            <form
-              className="code-box-code-action"
-              action="https://codesandbox.io/api/v1/sandboxes/define"
-              method="POST"
-              target="_blank"
-              ref={codeSandboxIconRef}
-              onClick={() => {
-                track({ type: 'codesandbox', demo: asset.id });
-                codeSandboxIconRef.current?.submit();
-              }}
-            >
-              <input
-                type="hidden"
-                name="parameters"
-                value={compress(JSON.stringify(codesanboxPrefillConfig))}
-              />
-              <Tooltip title={<FormattedMessage id="app.demo.codesandbox" />}>
-                <CodeSandboxIcon className="code-box-codesandbox" />
-              </Tooltip>
-            </form>
-            <CodeBlockButton title={localizedTitle} dependencies={dependencies} jsx={jsx} />
-            <Tooltip title={<FormattedMessage id="app.demo.stackblitz" />}>
-              <span
-                className="code-box-code-action"
-                onClick={() => {
-                  track({ type: 'stackblitz', demo: asset.id });
-                  stackblitzSdk.openProject(stackblitzPrefillConfig, {
-                    openFile: [`demo.${suffix}`],
-                  });
-                }}
-              >
-                <ThunderboltOutlined
-                  className="code-box-stackblitz"
-                  style={{ transform: 'scale(1.2)' }}
-                />
-              </span>
-            </Tooltip>
-            <form
-              className="code-box-code-action"
-              action="https://codepen.io/pen/define"
-              method="POST"
-              target="_blank"
-              ref={codepenIconRef}
-              onClick={() => {
-                track({ type: 'codepen', demo: asset.id });
-                codepenIconRef.current?.submit();
-              }}
-            >
-              <ClientOnly>
-                <input type="hidden" name="data" value={JSON.stringify(codepenPrefillConfig)} />
-              </ClientOnly>
-              <Tooltip title={<FormattedMessage id="app.demo.codepen" />}>
-                <CodePenIcon className="code-box-codepen" />
-              </Tooltip>
-            </form>
-            <Tooltip title={<FormattedMessage id="app.demo.separate" />}>
-              <a
-                className="code-box-code-action"
-                aria-label="open in new tab"
-                target="_blank"
-                rel="noreferrer"
-                href={demoUrlWithTheme}
-              >
-                <ExternalLinkIcon className="code-box-separate" />
-              </a>
-            </Tooltip>
-            <Tooltip
-              title={<FormattedMessage id={`app.demo.code.${codeExpand ? 'hide' : 'show'}`} />}
-            >
-              <div className="code-expand-icon code-box-code-action">
-                <img
-                  alt="expand code"
-                  src={
-                    theme?.includes('dark')
-                      ? 'https://gw.alipayobjects.com/zos/antfincdn/btT3qDZn1U/wSAkBuJFbdxsosKKpqyq.svg'
-                      : 'https://gw.alipayobjects.com/zos/antfincdn/Z5c7kzvi30/expand.svg'
-                  }
-                  className={codeExpand ? 'code-expand-icon-hide' : 'code-expand-icon-show'}
-                  onClick={() => handleCodeExpand(asset.id)}
-                />
-                <img
-                  alt="expand code"
-                  src={
-                    theme?.includes('dark')
-                      ? 'https://gw.alipayobjects.com/zos/antfincdn/CjZPwcKUG3/OpROPHYqWmrMDBFMZtKF.svg'
-                      : 'https://gw.alipayobjects.com/zos/antfincdn/4zAaozCvUH/unexpand.svg'
-                  }
-                  className={codeExpand ? 'code-expand-icon-show' : 'code-expand-icon-hide'}
-                  onClick={() => handleCodeExpand(asset.id)}
-                />
-              </div>
-            </Tooltip>
-          </Flex>
+          <Actions
+            showOnlineUrl={showOnlineUrl}
+            docsOnlineUrl={docsOnlineUrl}
+            codesanboxPrefillConfig={codesanboxPrefillConfig}
+            codepenPrefillConfig={codepenPrefillConfig}
+            stackblitzPrefillConfig={stackblitzPrefillConfig}
+            assetId={asset.id}
+            localizedTitle={localizedTitle}
+            dependencies={dependencies}
+            jsx={jsx}
+            suffix={String(suffix)}
+            demoUrlWithTheme={demoUrlWithTheme}
+            theme={
+              Array.isArray(theme) ? (theme[0] ? String(theme[0]) : '') : theme ? String(theme) : ''
+            }
+            codeExpand={codeExpand}
+            onCodeExpand={() => setCodeExpand((prev) => !prev)}
+          />
         </section>
       )}
       {codeExpand && (
